requirements.txt
Flask==2.3.2
twilio==7.18.0
openai>=0.11.0
python-dotenv==1.0.0
